frontend_src/pyproject.toml
[tool.poetry]
name = "CPAoR Project"
version = "1.0.0"
description = ""
authors = ["CPAoR developers"]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
pandas = "^2.2.2"
plotly = "^5.19.0"
python-dotenv = "1.0.1"
streamlit = "^1.34.0"
tqdm = "^4.66.2"
matplotlib = "^3.8.0"
fiona = "^1.9.5"
geopandas = "^0.14.3"
openpyxl = "^3.1.2"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
